Restore missing functionality (bugfix)
The reorganization in 5c5b4eeb3e10f0d5dc61b55e07032eadd1dc5e7e made it
so that there should no longer be a public `copy_folder` (preferring the
new `copy_folders`), but the logic from the former was still needed as
part of implementing the latter (indirectly via `_copy_folders`). It was
therefore added back in as `_copy_folder`, without the decorator.