Reorganize
The steps of the overall process have been identified:

* Determine the rules (a "generator") for finding children of a node
* Use these rules to set up a (Python) generator that traverses the tree
* "Visit" the tree and apply an "action" at each node

The filesystem-specific logic concerns both first and last steps, so a
sub-package has been created for each and `filesystem.py` is (for now)
copied to each with appropriate minimal fixes.

The test code is also adapted to match, including importing any
filesystem-specific stuff from the appropriate copy.