Relocate rejection logic
Rather than having `process` detect the `_rejected` sentinel, it's now
forwarded to the traversal, which checks for it explicitly.

This way allows naive iteration logic (a `for` loop, which implicitly
`.send`s back `None` values) to work as expected (rather than filtering
all folders since a falsy value was received).

Also added a test and modified the existing test to consider fewer
files.