Simplify accumulation interface
Instead of expecting a separate accumulator function, accumulation is
done by the main action. (This makes it easier to do things like
counting nodes that were successfully processed by the main part of the
action.)

The inner workings of processing now maintain a one-element list to hold
the "accumulator" object. This allows accumulation logic to work either
by replacing the accumulator with a return value (for example, it could
be a simple integer), or by mutating it in-place (and then also
returning it, so that it isn't replaced with `None`).

Added tests demonstrating the concept.