Rewrite overview
The overview is now split into a tl;dr and a more detailed section. The
tl;dr simply shows an example of the kind of code replacement the
project makes possible. The details go over the design more clearly, by
starting with a description of tree-processing tasks in general and then
describing the support offered for each sub-task.