Improve test setup
Isolated temporary directories are now requested for testing, and tests
can operate on source file trees set up in the temporary directories
rather than operating on the code base itself.

Zip files are used to describe the desired contents of the test file
trees (the temporary folder created by Pytest itself is treated as the
root for the traversal) because this makes it easy to describe the setup
procedure and easy to set up tests that include empty folders.

For each named test setup, there is a zip file containing the file
hierarchy to test, and a corresponding TOML file that will be used to
describe the expected results of tests.