Add `Traversal` class
The `Traversal` is a new abstraction that represents the process of
traversing the tree. It wraps an iterable (in particular, from a
`traverse` function) and provides an ordinary iterator, but can also be
called to invoke the main processing logic.

Processing logic is refactored so that the Traverse class can share code
with it more neatly.

A decorator is added so that calling a traverse generator can create a
`Traversal` instead. This is transparent to code that already used the
generator directly, or called `process`.