Add `accumulator` decorator
A decorator is provided to convert ordinary actions or action pairs into
ones that update an accumulator. The accumulated result is returned in
place of the original (`rejected` is handled specially), such that the
return value from the last node processed represents the "total".

The `NodeError` class is relocated to avoid a circular import.