Add test case
The README now shows how to use `accumulator` with a pair of actions
instead of just as a conventional decorator. The first `accumulator`
example also properly shows the `import` of `operator.add` (although the
need for this will hopefully soon be eliminated), and the needed
`accumulator` import itself is added to both examples.

A corresponding test case is added.