Fix iteration order
In order for sorted folder and file names to be processed in order,
while still performing a depth-first traversal, it's necessary to
reverse folder names before putting them on the stack.

This is awkward and it would be better to reverse the names for internal
nodes in the sorting routine (after first figuring out a cleaner way to
offer sorting), so as not to incur the overhead (especially since *most*
traversals will likely *not* expect sorted output).