Refactor traversal code
The process is now more clearly organized into a "step" for each
internal node encountered, which yields that node and then handles
children if the node is not "rejected". Children are "organized" by
partitioning and then optionally sorting, before yielding leaves and
pushing (for the current depth-first traversal; or enqueueing for
breadth-first traversal) internal children.

This structure should be more convenient for having other sorts of
traversals share code.

Note that the internal nodes added to the stack are no longer reversed
for an unsorted traversal; therefore, sorting children *in the getter*
will no longer work as expected. Client code is expected to pass a
`sort_key` to the traversal instead. This interface is all-around more
coherent.