Remove presumed-unnecessary `iter()` call
The `process` function is not called directly in tests, but provided for
client code that prefers not to use the `@traversal` decorator. Since
the provided traversal should already be a generator, the call to `iter`
appears to be redundant. (If e.g. a sequence were passed instead, the
code would fail anyway at the `.send` call.)