Add support for breadth-first traversals
A separate `@traversal` is constructed for breadth-first order, using a
`collections.deque` for the stack and choosing which end to take
elements from. Several aliases are added, and the `filesystem`
convenience module wraps them all appropriately.

A simple test is added for breadth-first traversal. This looks
unimpressive, since for the current test data the breadth-first order is
the same as the depth-first order. However, it did catch a bug: contrary
to what was written in previous commit messages (and comments), elements
added to the deque should only be reversed for depth-first traversals.